#8 Switch different environments
