+++ main.go
@@ -1,8 +1,8 @@
 package main
 
 import (
-	"github.com/ybkuroki/go-webapp-sample/common"
 	"github.com/ybkuroki/go-webapp-sample/config"
+	"github.com/ybkuroki/go-webapp-sample/migration"
 	"github.com/ybkuroki/go-webapp-sample/repository"
 	"github.com/ybkuroki/go-webapp-sample/router"
 )
@@ -13,8 +13,8 @@ func main() {
 	repository.InitDB()
 	db := repository.GetDB()
 
-	// TODO: switch the following processing by environment
-	common.InitMasterData()
+	migration.CreateDatabase(config.GetConfig())
+	migration.InitMasterData(config.GetConfig())
 
 	router := router.Init()
 	router.Start(":8080")
